convert rss platform use case to general controller
