+++ main.py
@@ -4,7 +4,7 @@ Main app file, all api route are declared there
 from flask import Flask, jsonify, request
 from flask_cors import CORS
 from application.use_cases.krebon_sec import KrebonSec
-from application.use_cases.dark_reading import DarkReading
+from application.interfaces.controllers.platform_rss_controller import PlatformRssController
 from application.use_cases.hacker_news import HackerNews
 from application.use_cases.threat_post import ThreatPost
 from application.use_cases.security_week import SecurityWeek
